+++ modules/state/outputs.tf
@@ -6,5 +6,8 @@ output "state_buckets" {
   value = { for scope, bucket in google_storage_bucket.state : scope => bucket.name }
 }
 output "state_replica_buckets" {
-  value = { for scope, bucket in google_storage_bucket.replica : scope => bucket.name }
+  value = { for scope, bucket in google_storage_bucket.replica_us : scope => bucket.name }
+}
+output "legacy_state_replica_buckets" {
+  value = { for scope, bucket in google_storage_bucket.legacy_replica : scope => bucket.name }
 }